13.08.2019

+++ src/server/game/Globals/AreaTriggerDataStore.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2018 TrinityCore <https://www.trinitycore.org/>
+ * Copyright (C) 2008-2019 TrinityCore <https://www.trinitycore.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -30,7 +30,7 @@ struct AreaTriggerData
     float position_y;
     float position_z;
     uint32 map_id;
-    uint32 spawn_mask;
+    uint64 spawn_mask;
     uint32 scriptId;
     ObjectGuid::LowType guid;
 };
